+++ src/types/Branch.ts
@@ -1,0 +1,8 @@
+import {City} from "@/types/City";
+
+export type Branch = {
+    calling_code: string
+    country: string
+    currency_code: string
+    cities: City[]
+}