+++ src/types/City.ts
@@ -1,0 +1,7 @@
+export type City = {
+    city: string
+    location: {
+        lat: number
+        lng: number
+    }
+}