+++ CHANGELOG.md
@@ -9,10 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Release Notes for 1.4.x
 
+## [1.4.2(2024-08-10)](https://github.com/DevAlexandreCR/admin-driver/compare/1.4.2...1.4.1)
+
+### Added
+- Add new implementation and view for see who created such services. [#136](https://github.com/DevAlexandreCR/gorda-admin/pull/136)
+- An Admin can Enable-Disable Drivers. [#135](https://github.com/DevAlexandreCR/gorda-admin/pull/135)
+
+### Fixed
+- Error when editing a user. [#142](https://github.com/DevAlexandreCR/gorda-admin/pull/142)
+
 ## [1.4.1(2024-07-24)](https://github.com/DevAlexandreCR/admin-driver/compare/1.4.1...1.4.0)
 
 ### Fixed
-
 - FIx query to load chats. [#137](https://github.com/DevAlexandreCR/gorda-admin/pull/141)
 
 ## [1.4.0(2024-07-23)](https://github.com/DevAlexandreCR/admin-driver/compare/1.4.0...1.3.5)